ping: Add back Data header
The IP header may contain an Options field after the Destination
Address.

Originally, when ping printed the IP header, it had Data in the title
row.  This was removed in ef9e6dc7eebe9830511602904d3ef5218d964080.
NetBSD has also removed it. It is present in OpenBSD and Linux.

Since a long time has passed, let's make Data visible only when there
are options.

Before (with options)
    Vr HL TOS  Len   ID Flg  off TTL Pro  cks      Src      Dst
     4  f  00 007c 0001   0 0000  40  01 d868 192.0.2.1  192.0.2.2 01...

After (with options)
    Vr HL TOS  Len   ID Flg  off TTL Pro  cks      Src      Dst Data
     4  f  00 007c 0001   0 0000  40  01 d868 192.0.2.1  192.0.2.2 01...

NB: The alignment may be fixed in a future commit.
NB: The Data title may change to Opts in a future commit.