ping: Pad Src and Dst headers
Initially, the Src/Dst IP addresses were printed in hex notation (%08x),
which will always be 8-characters wide.  Now we print them in
dot-decimal notation.  Depending on the IP address length, the Src and
Dst headers may require a different padding.

Use the source and destination IP lengths as padding for the headers.

Before (with options)
    Vr HL TOS  Len   ID Flg  off TTL Pro  cks      Src      Dst Data
     4  f  00 007c 0001   0 0000  40  01 d868 192.0.2.1  192.0.2.2 01...

After (with options)
    Vr HL TOS  Len   ID Flg  off TTL Pro  cks       Src        Dst Data
     4  f  00 007c 0001   0 0000  40  01 d868 192.0.2.1  192.0.2.2 01...

NB: Everything else in the header is in hex notation.
NB: GNU Inetutils has opted to use tabs to delimit these fields, but
    those are imprecise.