ping: Fix the spacing between the time stamp and cp/dp
When an echo reply packet is received, the data is compared with the
sent data.  When a wrong byte is detected the command displays a report
with the differences.

The first row (the first 8-bytes of data after the ICMP header) should
include the time stamp (if data is at least 8-bytes), this value is not
taken into consideration for the comparison.  The remaining rows
represent the data (padded pattern) received/sent, with each byte being
compared for differences.

Add an extra space after cp:/dp:.

It is worth noting that Darwin has opted to display everything as a
single block of data, however, we prefer to keep things as they are, as
it provides better visual representation of what is actually being
compared.

Before:
    cp:99  0  0  c  1  5  c  0
    	ab cd ab cd ab cd ab cd ab cd ab cd ab cd ab cd
    	...

After:
    cp: 99  0  0  c  1  5  c  0
    	ab cd ab cd ab cd ab cd ab cd ab cd ab cd ab cd
    	...