DROP ping: Clean up IP header
Use an updated version of Muuss' ping.c that never made it into BSD as
the inspiration for the IP header title and positioning.

All headers start with an uppercase letter and are left-aligned (the
Options header is already left-aligned).

Before:
    Vr HL TOS  Len   ID Flg  off TTL Pro  cks       Src        Dst
     4  5  00 0054 0001   0 0000  40  01 f6a4 192.0.2.1  192.0.2.2

After:
    Vr HL TOS Len  ID   Flg Off  TTL Pro Cksm Src       Dst
    4  5  00  0054 0001 0   0000 40  01  f6a4 192.0.2.1 192.0.2.2

NB: Do you prefer numbers right-aligned? maybe just change the titles?