ping_tests: Introduce a script to test ping
This script is largely inspired by injection.py, but it has been
extended to test most of ping.c.

This commit simply replaces the current injection.py with pinger.py.

It will provide the basis for series of fixes.

There is one test that is wrongly passing.  Let's mark it as XXX for
now, it will be fixed in future commits.

NB: The idea is just to get a review of the python/Scapy script itself.
And to use it as a drop-in replacement.