Anchor the functionality to -A
But -A needs to be fixed:

pinger() was always called, even if preload == 0.
This change was introduced in 32af342f5859a30002bdbfd3f82b1e1390b95fc9
which fixed two issues:
1. The first ping was sent undelayed. XXX
2. More than -c packets could be sent. This part is left untouched.

XXX investigate the timestamp, as gettimeofday() was called right after
it.

nmissedmax was introduced in 25107197363eae6ddfa026b308cd211196c20a1a
to account for a bug in -A...
redefine nmissedmas to be the max value of ntransmitted - nreceived.

XXX investigate the extra - 1.