ping: Add population standard deviation to SIGINFO
In this first iteration, let's duplicate the code from the final
statistics section to SIGINFO's.

- Avoid a potential square root of a negative number calculation
  (related to D38480)

I have a branch where it unifies the code onto a single `summary()`
function, however it changes a lot more of the code that I would like
and have tests for at the moment.  Besides, the code ends up being
almost identical to `ping6.c`'s.  Let's leave the unification for later
in the ping/ping6 series.

A few cosmetic things discovered from this exercise are:

- ping's SIGINFO prints out to stderr (ping6 to stdout)
- ping/ping6 should just become ping when displayed to the user
- ping uses the expression "somebody's printing up packets!", I prefer
  ping6's "somebody's duplicating packets!" when a duplicate packet is
  received
- ping calls the population standard deviation "stddev", while ping6
  calls it "std-dev".  I'm OK either way, but it should be the same for
  both