pkt-gen: Fix compilation on some 32-bit platforms
The type of tv_sec is time_t, which is 32-bit on some platforms.
The type of tv_nsec is long (long long only since C23).

This was discovered when building the port net/pkt-gen without any
patches, by ports fallout.

Cast to an intmax_t to avoid the warning/error, given -Werror (NB: The
original patch in the net/pkt-gen port used a long long, we opted for an
intmax_t instead).

See the aftermath of bug report #270440 for further details.