ping: Avoid reporting NaNs
Avoid calculating the square root of a negative number.

----

I'm not sure if this is the fix, as I could not reproduce it locally,
however some Jenkins tests are failing:

- https://ci.freebsd.org/job/FreeBSD-main-aarch64-test/1096/testReport/junit/sbin.ping/ping_test/ping6_c1_s8_t1/
- https://ci.freebsd.org/job/FreeBSD-main-aarch64-test/1096/testReport/junit/sbin.ping/ping_test/ping_c1_s56_t1_S127/
- https://ci.freebsd.org/job/FreeBSD-main-aarch64-test/1096/testReport/junit/sbin.ping/ping_test/ping_c1_s56_t1/

Is it possible to trigger a CI run without committing?

I plan on retaking my previous work on ping/ping6 soon-ish.