MAC/do: Better parsing for IDs (strtoui_strict())
Introduce strtoui_strict(), which signals an error on overflow contrary
to the in-kernel strto*() family of functions which have no 'errno' to
set and thus do not allow callers to distinguish a genuine maximum value
on input and overflow.

It is built on top of strtoq() and the 'quad_t' type in order to achieve
this distinction and also to still support negative inputs with the
usual meaning for these functions.  See the introduced comments for more
details.

Use strtoui_strict() to read IDs instead of strtol().

Reviewed by:    bapt
Differential Revision:  https://reviews.freebsd.org/D47614