queue: New debug macros for STAILQ
QMD_STAILQ_CHECK_TAIL() checks that the tail pointed by 'head' does not
have a next element.

QMD_STAILQ_CHECK_EMPTY() is a check for empty STAILQ, where heads's
'stqh_last' field must point to the 'stqh_first' one (which must be
NULL, but this has to be checked beforehand with STAILQ_EMPTY()).