crsetgroups_locked(): Now expects the input groups array sorted
This is in preparation for further commits introducing the setcred()
system call, which will need to sort groups early for MAC checks.

Not sorting an input array under a lock is also better practice (since
the lock is not needed at all), although in this case in normal
circumstances this is unlikely to make any visible difference.

To avoid disturbing all callers of crsetgroups(), the latter now sorts
the groups array before calling crsetgroups_locked() with qsort().  The
other caller of crsetgroups_locked(), namely kern_setgroups(), has been
adapted as well.

In passing, using qsort() instead of the old open-coded insertion sort
fixes the performance concern about the old insertion sort when using
a large number of groups.  Also, our qsort() falls back to insertion
sort for small arrays and in case the array is likely to be mostly
sorted, so this souldn't cause concerns for the common small number of
groups case either.