WIP nuageinit: Introduce sysrc_f()
Add a function that invokes `sysrc -f`, the `-f` flag is needed, when
faking the root directory, as `-R` cannot be used because it chroots and
expects a shell inside.  Otherwise, it can be omitted.

This function is slow, and we are investigating the possibility of
replacing it with a simple one.  Its only purpose, at the moment, is to
achieve idempotency.

It will be used latter to replace echoing to files, that can pollute
them with duplicate data.