git-arc: Improve 'list' when a review exists
When a review exists in Phabricator, git-arc matches the first line of
the commit log with the title of the review.  If there is a match, it
grabs the differential ID to query:

    echo '{"names":["D12345"]}' | arc call-conduit -- phid.lookup

and gets the "status" (e.g., open, closed) as reported by the API
instead of displaying the output similar to the one produced by:

    arc list --ansi

which appropriately shows a more accurate colored status (e.g.,
Accepted, Changes Planned, etc.)

Improve this by querying differential.revision.search, and
filtering the '.response.data[0].fields.status' response.

Also, bump the width to accommodate for "Changes Planned", which
contains 15 characters