BUG: Fix clean target when cross-building
When cross-building, either on macOS or Linux, make clean
fails due to not setting the proper machine/arch.

Evidently this is not the appropriate fix, however when trying to trace
the origin, I quickly got lost.  Maybe it is just easier to point it
here, so someone can really fix it, or further study the intricacies of
the FreeBSD build system.