nanobsd: Add a provisional populate_part function
Add a _populate_part(ition) function that mimics the current
populate_slice.  Note however, that this function is not
backward-compatible with populate_slice, hence the different name.  A
"_" is prepended to signal that it still experimental.

It can be used to populate the /cfg and /data partition using makefs(8).

Initially not wired.

TODO:

- Related: https://github.com/freebsd/poudriere/issues/1181
  With this approach, an mtree file will be required if a mix of
  different users/groups is desired.