blacklist: Adapt NetBSD's probes and changes to OpenSSH
Switch to using NetBSD's probes, there is no reason to have these in
different places.

Also, exclusively use BLACKLIST_AUTH_FAIL (BLACKLIST_BAD_USER was
recently changed upstream to count as 2 srtikes, and it can be
considered a POLA violation).

THESE CHANGE ARE CURRENTLY BEING TESTED IN THE WILD.

WIP:
Notice that NetBSD switched to using vsyslog_ss(3)[1], as a
countermeasure to [2].

[1]: https://github.com/NetBSD/src/commit/01636416235e7bcbfc829d372cfadd1cfd5494f8
[2]: https://www.freebsd.org/security/advisories/FreeBSD-SA-24:08.openssh.asc