blocklist: Rename blacklist to blocklist
Upstream has renamed blacklist to blocklist.

All the files have been renamed, the old library names were kept for
backward compatibility.

Utilities blacklistctl and blacklistd have been renamed to blocklistctl
and blocklistd respectively.

Configuration files with the old name will continue to work, but will
throw a warning, asking the user to change to the new naming.  Firewall
configurations with the old naming will continue to work, but will throw
a warning, asking the user to change the names (e.g., anchors or
sentinel files).

Manual pages only exist with the new naming.

The naming change will also be used to introduce two breaking changes
from upstream: one that requires that the db file be recreated; the
other that changes the weights in the penalty enum, now counting each
invalid attempt as two towards the maximum number of allowed fails.