bsdinstall: Harden reading the kernel message buffer
When choosing the option to harden reading the kernel message buffer by
an unprivileged user in bsdinstall, access to the kernel message buffer
is restricted via it sysctl (security.bsd.unprivileged_read_msgbuf=0).

In order for this setting to be effective, access to this data via
system logs must also be restricted.  Add the necessary hooks to
restrict access to the log files:

  - /var/log/messages
  - /var/run/dmesg.boot ($dmesg_file)

PR:		272552