release: packages: Add a dependency on geom to periodic
The periodic package provides a system for scheduling regular system
maintenance tasks.  This package also provides the base periodic tasks
for the base system.

One of those periodic tasks is to backup disk partition layouts
(221.backup-gpart), which is enabled by default.  This task is included
in the geom package.

This commit adds a dependency on the geom package to the periodic
package.