blocklist: Minimize differences with upstream
Commit 4d56eb007b18 ("blocklist: Revert upstream commit ddf6d71")
manually reverted an upstream change that affects how FreeBSD users
expect the nfail counter in blocklistd.conf to work.

Attempt to reduce the differences with upstream by re-implementing the
reverted upstream commit but as a one failure count.

Note that this will add functionality to BLOCKLIST_BAD_USER, which was
previously not implemented.  At this point, there are no probes in the
base distribution, nor in the ports tree that make use of this macro.