locale: monetdef: Add a callback to intsepbyspace
Add a callback to int_{n,p}_sep_by_space to insert a space between the
international currency symbol and the monetary quantity for en_IE and
en_US locales, matching common documentation practices[1] and the Open
Group's strfmon[2] manual page.  This aligns the output with typical
formatting conventions where a space is placed between the international
currency symbol and the amount for these locales.

The change affects only the formatting behavior of strfmon(3), as it is
the only function currently using int_{n,p}_sep_by_space.

[1]: https://style-guide.europa.eu/en/content/-/isg/topic?identifier=7.3.3-rules-for-expressing-monetary-units
[2]: https://pubs.opengroup.org/onlinepubs/9799919799/functions/strfmon.html