WIP Fix locales/man page
Use the SPDX license identifier.

Clarify that the "+" flag does not necessarily imply that a positive
sign is appended to the formatted string.

Extend the example to consider returned errors.

This commit also effectively reverts b211af11e213 ("strfmon(3) does not
correctly handle multibyte characters in the format string.").
Note that we are not claiming to be bug-free, just that there are no
known bugs.