df: ceil() the percentage value
"The percentage value shall be expressed as a positive integer, with any
fractional result causing it to be rounded [up] to the next highest
integer."
-- https://pubs.opengroup.org/onlinepubs/9799919799/utilities/df.html

Use a hand-rolled ceil() equivalent (as long as it doesn't overflow), to
avoid using math.h in our percent() function.

Fix the tests accordingly, they now match the percentage values on
NetBSD.

Also, use the same percent() function for the calculation of used
inodes.

Update the percentage values in the manual page as well.