makefs: Define HAVE_STRUCT_STAT_{BIRTHTIME,_ST_MTIMENSEC}
Define HAVE_STRUCT_STAT_BIRTHTIME and HAVE_STRUCT_STAT_ST_MTIMENSEC in
the main header file.

FreeBSD supports both.  For cross-builds, on Linux it is defined as 0,
and on macOS it is defined as 1 in their respective nbtool_config.h
header files.

Rearrange #if directives accordingly.

Tests marked with "XXX" will be addressed (or at least discussed)
separately.